% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipponen.R
\name{calculate_hrv_rmssd}
\alias{calculate_hrv_rmssd}
\title{Calculate RMSSD with Artefact Correction}
\usage{
calculate_hrv_rmssd(data, ...)
}
\arguments{
\item{data}{A \code{dplyr::tibble} with a 'time' column (RR intervals in ms).}

\item{...}{Arguments to be passed to \code{detect_hrv_artefacts} (e.g.,
\code{alpha}, \code{c1}, \code{c2}).}
}
\value{
A list with the following components
\itemize{
\item \code{rmssd_values}: The RMSSD value calculated from the \emph{corrected} RR
intervals.
\item \code{classified_data}: The data with beat classification
\item \code{corrected_data}: The data with beat correction
}
}
\description{
A convenience function that combines \code{detect_hrv_artefacts} and
\code{correct_hrv_beats} to perform both artefact detection/classification and
correction in a single step, and then calculate the RMSSD.
}
