% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{extract_rr_intervals}
\alias{extract_rr_intervals}
\title{Extract RR intervals from FIT file}
\usage{
extract_rr_intervals(
  fit_object,
  filter_factor,
  laying_time = 180,
  standing_time = 180
)
}
\arguments{
\item{fit_object}{FITfileR object}

\item{filter_factor}{Used to filter RR values that are 1.x times smaller or
larger than the last one as suspected false readings.}
}
\value{
List of laying and standing RR intervals in ms
}
\description{
Extract RR intervals from FIT file
}
\keyword{internal}
