% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_training_recommendations}
\alias{generate_training_recommendations}
\title{Generate training recommendations based on neural recovery score}
\usage{
generate_training_recommendations(score, primary_type = "BJJ")
}
\arguments{
\item{score}{Neural recovery score}

\item{primary_type}{Primary training type ("BJJ" or "Strength")}
}
\value{
List of training recommendations
}
\description{
Generate training recommendations based on neural recovery score
}
