% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv.R
\name{get_HR}
\alias{get_HR}
\title{Extract HR data from \code{FitFile} object}
\usage{
get_HR(fit_object)
}
\arguments{
\item{fit_object}{An object of class \code{FitFile}, eg.
from \code{\link[FITfileR::readFitFile]{FITfileR::readFitFile}}}
}
\value{
A vector containing up to the first three minutes of heart rate data.
}
\description{
Extract HR data from \code{FitFile} object
}
