% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv.R
\name{get_RR}
\alias{get_RR}
\title{Calculate RR from HRV data.}
\usage{
get_RR(HRV, filter_factor)
}
\arguments{
\item{HRV}{A vector of HRV data (e.g. from \link{get_HRV})}

\item{filter_factor}{Used to filter RR values that are 1.x times smaller or
larger than the last one as suspected false readings.}
}
\value{
A named list containing the full RR data and suspected RR data for
  standing and lying down.
}
\description{
Calculate RR from HRV data.
}
