% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv.R
\name{hrv_metrics}
\alias{hrv_metrics}
\title{Calculate HRV summaries for an orthostatic test}
\usage{
hrv_metrics(fit_object, filter_factor = 0.25)
}
\arguments{
\item{fit_object}{An object of class \code{FitFile}, eg.
from \code{\link[FITfileR::readFitFile]{FITfileR::readFitFile}}}

\item{filter_factor}{Used to filter RR values that are 1.x times smaller or
larger than the last one as suspected false readings.}
}
\value{
A named list of HRV summarie metrics.
}
\description{
Calculate HRV summaries for an orthostatic test
}
