% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hrv_trend_plot}
\alias{hrv_trend_plot}
\title{Trend plots of HRV summaries for orthostatic tests}
\usage{
hrv_trend_plot(metrics, just_rssme = FALSE)
}
\arguments{
\item{metrics}{A tibble or data frame containing HRV metrics. Output of
\link{process_fit_directory}}

\item{just_rssme}{Logical indicating whether to focus on RMSSD and resting HR
metrics. Default: FALSE}
}
\value{
A ggplot2 plot object
}
\description{
Creates trend plots of HRV metrics, showing both daily values and moving averages.
}
