% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvester-package.R
\docType{package}
\name{hrvester-package}
\alias{hrvester}
\alias{hrvester-package}
\title{hrvester: Simple Heart Rate Variability Analysis}
\description{
Provides tools for analyzing Heart Rate Variability (HRV) from FIT files.
Focuses on orthostatic test analysis (laying to standing measurements).
}
\section{Main Functions}{

\itemize{
\item process_fit_directory(): Process multiple FIT files with caching
\item process_fit_file(): Process a single FIT file
\item plot_hrv_trend(): Create trend plots of HRV metrics
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sims1253/hrvester}
  \item \url{https://sims1253.github.io/hrvester/}
  \item Report bugs at \url{https://github.com/sims1253/hrvester/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Scholz \email{dev.scholz@mailbox.org}

Authors:
\itemize{
  \item Maximilian Scholz \email{dev.scholz@mailbox.org}
}

}
\keyword{internal}
