% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hrv_trends}
\alias{plot_hrv_trends}
\title{Plot HRV trends}
\usage{
plot_hrv_trends(data)
}
\arguments{
\item{data}{Data frame containing HRV metrics}
}
\value{
A ggplot2 plot object
}
\description{
Creates a trend plot of RMSSD values with a moving average and normal range
shading.
}
