% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_processing.R
\name{rr_validate_measure_artifacts}
\alias{rr_validate_measure_artifacts}
\title{RR validation that flags intervals of 65.535 as invalid}
\usage{
rr_validate_measure_artifacts(
  rr_segment,
  is_valid = rep(TRUE, length(rr_segment))
)
}
\arguments{
\item{rr_segment}{A numeric vector of RR intervals (in milliseconds).}

\item{is_valid}{A logical vector indicating the validity of each RR interval. Defaults to TRUE.}
}
\value{
A list containing two elements:
\item{is_valid}{A logical vector of the same length as \code{rr_segment}, where
\code{TRUE} indicates a valid RR interval and \code{FALSE} indicates an invalid
interval.}
\item{cleaned_rr}{A numeric vector containing only the valid RR intervals.}
}
\description{
When taking the raw RR output from a Garmin watch, most of the entries are 65.535, which is
just an empty measurement, as no new heartbeat has been detected yet.
}
