% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate input parameters}
\usage{
validate_inputs(dir_path, cache_file, filter_factor, clear_cache)
}
\arguments{
\item{dir_path}{Directory path}

\item{cache_file}{Cache file path}

\item{filter_factor}{Filter factor value}

\item{clear_cache}{Clear cache flag}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Internal function to validate input parameters for process_fit_directory
}
\keyword{internal}
