% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calculate_hrr}
\alias{calculate_hrr}
\title{Calculate Heart Rate Recovery metrics}
\usage{
calculate_hrr(standing_hr, baseline_hr)
}
\arguments{
\item{standing_hr}{Vector of heart rate values after standing}

\item{baseline_hr}{Resting heart rate before standing}
}
\value{
List of HRR metrics
}
\description{
Calculate Heart Rate Recovery metrics
}
