% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_hrr}
\alias{calculate_hrr}
\title{Calculate Heart Rate Recovery metrics}
\usage{
calculate_hrr(standing_hr, baseline_hr)
}
\arguments{
\item{standing_hr}{Vector of heart rate values after standing}

\item{baseline_hr}{Resting heart rate before standing}
}
\value{
List containing recovery metrics:
\item{hrr_60s}{Absolute recovery in 60 seconds}
\item{hrr_relative}{Relative recovery percentage}
\item{orthostatic_rise}{Initial HR increase}
}
\description{
Computes various metrics related to heart rate recovery
}
