% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calculate_moving_averages}
\alias{calculate_moving_averages}
\title{Calculate all moving averages}
\usage{
calculate_moving_averages(data, window_size = 7, min_fraction = 0.7)
}
\arguments{
\item{data}{Dataframe of HRV measurements}

\item{window_size}{Number of days for moving average}

\item{min_fraction}{Minimum fraction of data required}
}
\value{
Dataframe with added moving averages
}
\description{
Adds moving averages and relative changes to HRV data
}
