% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_neural_recovery}
\alias{calculate_neural_recovery}
\title{Calculate full neural recovery score based on BioForce HRV methodology}
\usage{
calculate_neural_recovery(data, window_size = 7)
}
\arguments{
\item{data}{Dataframe containing HRV measurements}
}
\value{
Dataframe with added neural recovery scores
}
\description{
Calculate full neural recovery score based on BioForce HRV methodology
}
