% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipponen.R
\name{calculate_rmssd_orthostatic_enhanced}
\alias{calculate_rmssd_orthostatic_enhanced}
\title{Calculate RMSSD with Phase-Specific Artefact Correction (Orthostatic Test)}
\usage{
calculate_rmssd_orthostatic_enhanced(
  data,
  secondary_threshold_percent_lying = 20,
  secondary_threshold_percent_standing = 20,
  transition_exclusion_time = 45,
  initial_stabilization_time = 60,
  min_segment_length = 30,
  min_segment_beats = 30,
  ...
)
}
\arguments{
\item{data}{A \code{dplyr::tibble} with a 'time' column (RR intervals).  Assumes
the first 3 minutes are lying and the next 3 minutes are standing.}

\item{secondary_threshold_percent_lying}{Numeric, percentage change
threshold for secondary filtering during the lying phase (default: 20).}

\item{secondary_threshold_percent_standing}{Numeric, percentage change
threshold for secondary filtering during the standing phase (default: 20).}

\item{transition_exclusion_time}{Numeric, duration (in seconds) to exclude
around the transition (default: 45).  Increased default.}

\item{initial_stabilization_time}{Numeric, duration (in seconds) to exclude
at the beginning of the lying phase (default: 60).}

\item{min_segment_length}{Numeric, minimum segment length in seconds
(default: 30).}

\item{min_segment_beats}{Integer, minimum number of beats in a segment
(default: 30)}

\item{...}{other parameters passed to \code{detect_hrv_artefacts}}
}
\value{
A list containing:
\itemize{
\item \code{rmssd_lying}: RMSSD for the lying phase (or NA).
\item \code{rmssd_standing}: RMSSD for the standing phase (or NA).
\item \code{rmssd_values}: A numeric vector of all segment RMSSD values.
\item \code{segment_lengths}: A numeric vector of segment lengths.
\item \code{segment_beat_counts}: An integer vector of segment beat counts.
\item \code{aggregated_rmssd}: Aggregated RMSSD (mean of a ll segments).
\item \code{n_segments}: Number of segments
}
}
\description{
Calculate RMSSD with Phase-Specific Artefact Correction (Orthostatic Test)
}
