% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calculate_robust_ma}
\alias{calculate_robust_ma}
\title{Calculate rolling mean with minimum data requirement}
\usage{
calculate_robust_ma(x, window = 7, min_fraction = 0.7)
}
\arguments{
\item{x}{Numeric vector}

\item{window}{Size of rolling window}

\item{min_fraction}{Minimum fraction of non-NA values required (0-1)}
}
\description{
Calculate rolling mean with minimum data requirement
}
