% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_daily_report}
\alias{generate_daily_report}
\title{Generate daily HRV report}
\usage{
generate_daily_report(data)
}
\arguments{
\item{data}{HRV data frame}
}
\value{
Formatted report
}
\description{
Generate daily HRV report
}
