% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv.R
\name{hrv_plot}
\alias{hrv_plot}
\title{Plot HRV summaries for a single orthostatic test fit file.}
\usage{
hrv_plot(fit_file_path, base = "RR", filter_factor = 0.25)
}
\arguments{
\item{fit_file_path}{Path to the fit file}

\item{base}{"HR" if x axis should be heart rate or "RR" of x axis should be RR.}

\item{filter_factor}{Used to filter RR values that are 1.x times smaller or
larger than the last one as suspected false readings.}
}
\value{
A ggplot2 plot object.
}
\description{
Plot HRV summaries for a single orthostatic test fit file.
}
