% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv.R
\name{hrv_trend_plot}
\alias{hrv_trend_plot}
\title{Trend plots of HRV summaries for orthostatic tests}
\usage{
hrv_trend_plot(fit_dir)
}
\arguments{
\item{fit_dir}{Path of directory containing fit files.}
}
\value{
A ggplot2 plot object.
}
\description{
Trend plots of HRV summaries for orthostatic tests
}
