% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_processing.R
\name{validate_validity}
\alias{validate_validity}
\title{Validate Validity Vector}
\usage{
validate_validity(is_valid, rr_segment)
}
\arguments{
\item{is_valid}{A logical vector indicating the validity of each RR interval.}

\item{rr_segment}{A numeric vector of RR intervals (in milliseconds).}
}
\description{
This function validates a logical vector indicating the validity of RR
intervals, ensuring it has the correct type, length, and contains no NA
values.
}
\keyword{internal}
